Add navigation between Home and Recipe page
